Sprint 4 Przygotowanie pod logowanie

+++ src/main/java/com/example/tsp_server/repository/LanguageRepository.java
@@ -7,9 +7,6 @@ import org.springframework.stereotype.Repository;
 import java.util.Optional;
 
 @Repository
-public interface LanguageRepository extends JpaRepository<Language, Integer> {
-    // Lista wszystkich języków jest już dostępna przez dziedziczenie z JpaRepository
-
-    // Dodanie metody do wyszukiwania języka po nazwie
+public interface LanguageRepository extends JpaRepository<Language, Long> {
     Optional<Language> findByName(String name);
 }
